+++ tool_packages/unique_deep_research/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.2] - 2025-10-10
+- Get website title for OpenAI agent
+- Bolding of message logs
+- Clarifying questions and research brief dependent on engine type
+
 ## [3.0.1] - 2025-10-08
 - Improved citation logic supporting internal search documents
 - Fixed bug in referencing of internal sources not giving the correct title of sources